Remove misleading TMI
Was rereading my old posts and noticed that this was a bit of medical
TMI which I'd rather not have on my blog. It's also just kind of
inaccurate - I was already in remission for my Crohn's disease by the
time of this post and the ongoing treatment doesn't have a significant
effect.